Setup the bootstrapping script and test docs
I'm using a Python script to do a simplified version of what `Cobble`
will actually do. This is because I want to write the actual
implementation of `Cobble` literatly, so that I can explain my thought
process and also have it test itself. The examples show the most basic
functionality needed in `Cobble`:
- Exporting code blocks to source code files
- Naming blocks
- Referencing blocks through their names and appending / prepending
  prefixes / suffixes at the start of every line
- Accessing blocks from other files