Setup the actual project
I setup the project this way because I want it to be somewhat like a
UNIX utility. That is why `Cobble` is actually just something that
parses a TOML IR that is generated by some other program, so that it
can be modular and support other document formats through the use of
other frontends and parsers.